Multi-machine refactor: unify MDU state struct + share IRQ callbacks
Mirror the GLUE-family state unification for MDU+RBV: replace the separate
iici_state_t and iisi_state_t with one mac030_mdu_state_t (in mdu_io.h) — a
superset where the IIsi uses egret + last_port_a and the IIci uses last_port_b
and leaves egret NULL.  iici/iisi also drop their identical scc_irq/via1_irq
and use the shared mac030_glue_scc_irq / mac030_glue_via1_irq (same IRQ bit
values).

Note on scope: a board-driven unified init is deliberately NOT applied to
MDU/OSS. The IIci and IIsi diverge substantially (Egret vs none, IICI vs V8
RBV variant, two-bank RAM, ADB-via-VIA1 vs ADB-via-Egret, slot-$B vs slot-$E
video) — a single init would need ~13 board hooks for two machines, the
"kitchen sink" the proposal's §7 warns against. iifx (OSS) is a family-of-one,
so board-init yields no dedup. The GLUE near-clone triplet remains the
canonical board-init case; MDU/OSS keep purpose-built inits.

Verified: headless + WASM build clean (exit code checked explicitly);
iici-boot, iisi-boot pass.
